Add IWindowManager APIs to get/save overlay windowing mode with WMS context
The overlay windowing mdoe is stored /data/system_ce/user_id, and it
needs very important permission, and it is not allowed by system_app. I
dont' want to change neverallow rules, so I just add new APIs to
leverage WindowManagerService's context to avoid this problem.

Also, this CL changes the rule to use overlay windowing mode. If the
overlay windowing mode is UNDEFINED, it will skip it and use windowing
mode stored in normal windowing mode file.

The next step is to add more permission checking for added APIs.

Change-Id: I695fad7948553b16701d25d2d8efbe6465939598
Signed-off-by: utzcoz <utzcoz@outlook.com>